requirements.txt
# Core dependencies
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0

# GUI and interactive features
streamlit>=1.20.0

# Performance monitoring
psutil>=5.8.0

# Testing and development
pytest>=6.0.0
pytest-cov>=2.0.0

# Code quality
black>=21.0.0
flake8>=3.8.0
mypy>=0.800

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0 